Solución de errores general del projecto + gitignore ahora está en el root
