+++ .gitignore
@@ -1,3 +1,7 @@
+node_modules
+.env
+public
+
 # Logs
 logs
 *.log
@@ -7,7 +11,6 @@ yarn-error.log*
 pnpm-debug.log*
 lerna-debug.log*
 
-node_modules
 dist
 dist-ssr
 *.local
