+++ client/vite-project/src/pages/DoctorConsults.jsx
@@ -1,14 +1,14 @@
 // src/pages/DoctorConsults.jsx
 import React from 'react';
-import ConsultList from '../components/ConsultList';
+import ConsultList from './ConsultList';
 
 const DoctorConsults = () => {
-    return (
-        <div>
-            <h1>Consultas para Doctores</h1>
-            <ConsultList />
-        </div>
-    );
+  return (
+    <div>
+      <h1>Consultas para Doctores</h1>
+      <ConsultList />
+    </div>
+  );
 };
 
 export default DoctorConsults;
