+++ server/src/controllers/consults/index.js
@@ -4,7 +4,7 @@ import listConsultsController from './listConsultsController.js';
 import getConsultByIdController from './getConsultByIdController.js';
 import voteConsultController from './voteConsultController.js';
 import deleteConsultController from './deleteConsultController.js';
-import replyUserController from './replyConsultController.js';
+import replyConsultController from './replyConsultController.js';
 import deleteReplyConsultId from './deleteReplyConsultId.js';
 
 export {
@@ -13,6 +13,6 @@ export {
     getConsultByIdController,
     voteConsultController,
     deleteConsultController,
-    replyUserController,
+    replyConsultController,
     deleteReplyConsultId,
 };
