+++ server/src/controllers/consults/listConsultsController.js
@@ -5,7 +5,6 @@ import getPool from '../../db/getPool.js';
 const listconsultsController = async (req, res, next) => {
     try {
         // Obtenemos los query params necesarios.
-        let { place, author } = req.query;
 
         // Obtenemos una conexión con la base de datos.
         const pool = await getPool();
@@ -21,13 +20,9 @@ const listconsultsController = async (req, res, next) => {
                 c.createdAt
             FROM consults c
             INNER JOIN users u ON u.id = c.userId
-            WHERE u.username LIKE ?
-        `,
-            // Si "place" o "author" es undefined establecemos un string vacío. De lo contrario no
-            // figurará ninguna entrada como resultado.
-            [`%${place || ''}%`, `%${author || ''}%`]
+            
+        `
         );
-
 
         // Enviamos una respuesta al cliente.
         res.send({
